chore: add some helpful options to the shell SSH command
The new options prevent SSH from editing the user's normal SSH host key
file and hide the warning that a host key is being added (which it would
otherwise still print even though it's not editing a real
file). Changing the overall log level is perhaps a large hammer, but I
don't think there are a lot of other warning-level messages likely to be
shown.