wip: systemd socket activation for master-in-Docker
Docker evidently does not support passing a file descriptor into a
container, so using socket activation with a Docker container requires
passing the connection through a proxy program provided by systemd,
systemd-socket-proxyd.

By default, if the process inside the container takes any time to begin
listening on the port, the proxy will report connection failures until
the process is ready. The ExecStartPost line in the service prevents
that by causing systemd to not treat the service as started until it is
able to accept a connection.

[WIP note: All the values are just baked in since this is currently only
meant to be a demonstration that this technique is possible.]